Crate Made
There is now a crate, and the player can pick it up by running into it, then drop it as well in front of them. I have not been able to make the crate drop further away while the player turns around in varying directions, but the crate can be picked up and dropped while falling down instead of being teleported.