Mini Puzzles Made And HUD Added
I was not able to add a lot of what I wanted to add, I even had to get rid of a lot of the ideas I had, because I realized I am running out of time quicker than I had thought. The game does run and has all of the core mechanics I need with the clone and movement, and there is a basic HUD that can transition through the controls, main menu, etc.